requirements.txt
defusedxml==0.7.1
emoji==2.12.1
fonttools==4.53.1
fpdf2==2.7.9
Jinja2==3.1.4
Markdown==3.6
markdown2==2.5.0
MarkupSafe==2.1.5
packaging==24.1
pdfkit==1.0.0
pillow==10.4.0
PyPDF2==3.0.1
pytesseract==0.3.10
python-dotenv==1.0.1
PyYAML==6.0.1
typing_extensions==4.12.2
